Moving transaction address generation related functions to addresses.go
