+++ internal/wallet/core.go
@@ -236,25 +236,16 @@ func initializeWallet(seedPhrase string, pubPass []byte, privPass []byte, baseDi
 		log.Fatalf("Error deriving account key: %v", err)
 	}
 
-	// Print master fingerprint
-	masterFingerprint, err := getMasterFingerprint(rootKey)
-	if err != nil {
-		log.Fatalf("Error getting master fingerprint: %v", err)
-	}
-	log.Printf("Master Fingerprint: %x", masterFingerprint)
-
 	// Get xpub and zpub
-	xpub, err := getExtendedPubKey(accountKey, []byte{0x04, 0x88, 0xB2, 0x1E})
+	_, err = GetExtendedPubKey(accountKey, []byte{0x04, 0x88, 0xB2, 0x1E})
 	if err != nil {
 		log.Fatalf("Error getting xpub: %v", err)
 	}
-	log.Printf("xpub: %s", xpub)
 
-	zpub, err := getExtendedPubKey(accountKey, []byte{0x04, 0xB2, 0x47, 0x46})
+	_, err = GetExtendedPubKey(accountKey, []byte{0x04, 0xB2, 0x47, 0x46})
 	if err != nil {
 		log.Fatalf("Error getting zpub: %v", err)
 	}
-	log.Printf("zpub: %s", zpub)
 
 	log.Println("Initializing Neutrino chain service")
 	db, err := walletdb.Create("bdb", filepath.Join(neutrinoDBPath, "neutrino.db"), true, time.Second*60)
