+++ internal/wallet/utils.go
@@ -15,12 +15,7 @@ import (
 	"strings"
 	"time"
 
-	walletstatedb "github.com/Maphikza/btc-wallet-btcsuite.git/internal/database"
-	transaction "github.com/Maphikza/btc-wallet-btcsuite.git/lib/transaction"
-	"github.com/Maphikza/btc-wallet-btcsuite.git/lib/utils"
-	"github.com/btcsuite/btcd/btcutil"
 	"github.com/btcsuite/btcd/btcutil/hdkeychain"
-	"github.com/btcsuite/btcwallet/chain"
 	"github.com/btcsuite/btcwallet/wallet"
 	"github.com/joho/godotenv"
 	"golang.org/x/crypto/ripemd160"
@@ -55,7 +50,7 @@ func deriveKeyFromPath(rootKey *hdkeychain.ExtendedKey, path string) (*hdkeychai
 }
 
 // getMasterFingerprint calculates the master fingerprint from the root key
-func getMasterFingerprint(rootKey *hdkeychain.ExtendedKey) (uint32, error) {
+func GetMasterFingerprint(rootKey *hdkeychain.ExtendedKey) (uint32, error) {
 	pubKey, err := rootKey.ECPubKey()
 	if err != nil {
 		return 0, fmt.Errorf("failed to get public key from root key: %v", err)
@@ -77,7 +72,7 @@ func getMasterFingerprint(rootKey *hdkeychain.ExtendedKey) (uint32, error) {
 }
 
 // getExtendedPubKey converts an extended key to its string representation with the given version bytes
-func getExtendedPubKey(extendedKey *hdkeychain.ExtendedKey, version []byte) (string, error) {
+func GetExtendedPubKey(extendedKey *hdkeychain.ExtendedKey, version []byte) (string, error) {
 	neuteredKey, err := extendedKey.Neuter()
 	if err != nil {
 		return "", err
@@ -171,70 +166,6 @@ func gracefulShutdown() error {
 	return nil
 }
 
-func handleAddressGeneration(w *wallet.Wallet, chainClient *chain.NeutrinoClient, needsAddresses, freshWallet bool) error {
-	var numberofAddr int
-	if needsAddresses {
-		numberofAddr = 30
-		err := generateInitialAddresses(w, chainClient, numberofAddr)
-		if err != nil {
-			return fmt.Errorf("error generating initial addresses: %s", err)
-		}
-	} else if freshWallet {
-		numberofAddr = 1
-		err := generateInitialAddresses(w, chainClient, numberofAddr)
-		if err != nil {
-			return fmt.Errorf("error generating initial addresses: %s", err)
-		}
-	} else {
-		log.Println("Using existing addresses from the database")
-	}
-
-	return nil
-}
-
-func generateInitialAddresses(w *wallet.Wallet, chainClient *chain.NeutrinoClient, numAddresses int) error {
-	const maxRetries = 2
-	var receiveAddresses, changeAddresses []btcutil.Address
-	var err error
-
-	for i := 0; i < maxRetries; i++ {
-		log.Printf("Attempt %d: Generating initial addresses", i+1)
-		receiveAddresses, changeAddresses, err = utils.GenerateAndSaveAddresses(w, numAddresses)
-		if err == nil {
-			log.Printf("Successfully generated and saved %d receive addresses and %d change addresses", len(receiveAddresses), len(changeAddresses))
-			break
-		}
-		log.Printf("Error generating addresses: %v", err)
-		if i < maxRetries-1 {
-			log.Println("Retrying in 5 seconds...")
-			time.Sleep(5 * time.Second)
-		}
-	}
-
-	if err != nil {
-		log.Printf("Failed to generate addresses after %d attempts", maxRetries)
-		return fmt.Errorf("failed to generate addresses after %d attempts, with error: %s", maxRetries, err)
-	} else {
-		utils.PrintAddresses("Receive", receiveAddresses)
-		utils.PrintAddresses("Change", changeAddresses)
-	}
-
-	_, chainClientBestblock, err := chainClient.GetBestBlock()
-	if err != nil {
-		return fmt.Errorf("error getting chain client best block: %v", err)
-	} else {
-		log.Printf("Chain client best block: %d", chainClientBestblock)
-		err = walletstatedb.SetLastScannedBlockHeight(chainClientBestblock)
-		if err != nil {
-			return fmt.Errorf("error setting initial last scanned block height: %v", err)
-		} else {
-			log.Printf("Initial last scanned block height set to %d", chainClientBestblock)
-		}
-	}
-
-	return nil
-}
-
 func generateRandomPassphrase(length int) (string, error) {
 	bytes := make([]byte, length)
 	if _, err := rand.Read(bytes); err != nil {
@@ -249,28 +180,4 @@ func GetWalletBalance(w *wallet.Wallet) (int64, error) {
 		return 0, fmt.Errorf("error calculating balance: %v", err)
 	}
 	return int64(balance), nil
-}
-
-func EstimateTransactionSize(w *wallet.Wallet, spendAmount int64, recipientAddress string, feeRate int) (int, error) {
-	return transaction.HttpCalculateTransactionSize(w, spendAmount, recipientAddress, feeRate)
-}
-
-func GetTransactionHistory(w *wallet.Wallet, walletName string) ([]map[string]interface{}, error) {
-	transactions, err := w.ListAllTransactions()
-	if err != nil {
-		return nil, fmt.Errorf("error listing transactions: %v", err)
-	}
-
-	var result []map[string]interface{}
-
-	for _, tx := range transactions {
-		transaction := map[string]interface{}{
-			"txid":   tx.TxID,
-			"date":   time.Unix(tx.Time, 0).Format(time.RFC3339),
-			"amount": fmt.Sprintf("%.8f", tx.Amount),
-		}
-		result = append(result, transaction)
-	}
-
-	return result, nil
 }
